Added methods to multiple classes.
Added peek, size, and isEmpty method to productBacklog, but I will not
make productBacklog implement Iterable or Collection because that class
should not be used polymorphically.

Added model.command packge which includes functional interface Command,
one implementing class AddCommand, and a factory class CommandFactory so
that new commands can be created without worrying about the implementing
class, type arguments, or instance of Product Backlog passed as
arguments to the new commands.

Added forEachNext method to iterator to make it easier to apply consumer
to every element of an iteration. As such, I modified the return type of
the select method in ProductBacklog back to Iterator since Iterator can
use forEach.

I updated JUnit testing to accomidate to some of these changes.